update affected tests
