--- tests/component/_task_modules/channels/__init__.py
@@ -1,1 +1,0 @@
-"""Component tests for nidaqmx._task_modules.channels.*."""
