+++ tests/component/task/channels/__init__.py
@@ -1,0 +1,1 @@
+"""Component tests for nidaqmx.task.channels.*."""
