+++ tests/component/task/channels/test_ai_channel.py
@@ -4,7 +4,7 @@ from typing import List
 import pytest
 
 from nidaqmx import Task
-from nidaqmx._task_modules.channels import AIChannel
+from nidaqmx.task.channels import AIChannel
 from nidaqmx.constants import (
     AccelChargeSensitivityUnits,
     AccelSensitivityUnits,
