+++ tests/component/task/test_export_signals_properties.py
