+++ tests/component/task/test_in_stream_buffer_properties.py
