+++ tests/component/task/test_in_stream_read_properties.py
