+++ tests/component/task/test_out_stream_write_properties.py
