+++ tests/component/task/test_timing_properties.py
