+++ tests/component/task/test_triggers_properties.py
